+++ stdlib/tests/crypto/fri/channel.rs
@@ -1,8 +1,7 @@
 use processor::Digest;
 use test_utils::{
-    Felt, FieldElement, MerkleTreeVC, StarkField,
+    Felt, FieldElement, MerkleTreeVC,
     crypto::{BatchMerkleProof, ElementHasher, Hasher as HasherTrait, PartialMerkleTree},
-    math::fft,
     serde::DeserializationError,
 };
 use winter_fri::{FriProof, VerifierError};
@@ -69,7 +68,7 @@ where
         self.layer_commitments.drain(..).collect()
     }
 
-    pub fn read_remainder<const N: usize>(
+    pub fn read_remainder(
         &mut self,
         expected_commitment: &<H as HasherTrait>::Digest,
     ) -> Result<Vec<E>, VerifierError> {
@@ -77,11 +76,6 @@ where
         let commitment = H::hash_elements(&poly);
         assert_eq!(&commitment, expected_commitment);
 
-        // Compute remainder codeword corresponding to remainder polynomial
-        let twiddles = fft::get_twiddles(poly.len());
-        let remainder =
-            fft::evaluate_poly_with_offset(&poly, &twiddles, E::BaseField::GENERATOR, 8);
-
-        Ok(remainder)
+        Ok(poly)
     }
 }